Env .py and .rst files updated, closes #20
