+++ src/mlpro_int_gymnasium/envs/__init__.py
@@ -1,1 +1,2 @@
 from .multicartpole import *
+from .rubikscube2x2x2 import *
